+++ routes/modelIndex.js
@@ -88,17 +88,15 @@ var route = function (req, res, next) {
 
                 (cb) => {
 
-                    if (!req.linz.notifications.length) {
-                        return cb();
-                    }
-
-                    linz.api.views.renderPartial('notifications', { notifications: req.linz.notifications }, (err, notificationHtml) => {
+                    linz.api.views.renderNotifications(req, (err, notificationHtml) => {
 
                         if (err) {
-                            return res.send(err);
+                            return cb(err);
                         }
 
-                        data.notifications = notificationHtml;
+                        if (notificationHtml) {
+                            data.notifications = notificationHtml;
+                        }
 
                         return cb();