+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Fixed Linz not being able to find labels for embedded documents.
 - Fixed ckeditor for non `/admin` admin paths.
 - Fixed scoping of the multiselect within a documentarray.
+- Reworked some Linz internals to allow code using `linz.api.views.render` to take advantage of Linz's Notifications API. Also moved the render notifications functionality into the `linz.api.views` namespace and updated all code that rendered notifications to use this API function.
+- Moved Linz's `namespace` middleware into `middleware-public` so that it can be accessed in other applications using `linz.middleware.namespace`.
+- Moved Linz's `notifications` middleware into `middleware-public` so that it can be accessed in other applications using `linz.middleware.notifications`.
 
 ## v1.0.0-14.0.0 (14 November 2017)